requirements.txt
python-telegram-bot[job-queue]
transmission-rpc
routeros-api
python-dotenv